ASoC: adau1977: remove platform data
Remove platform data header file. Move MICBIAS
enum to local driver header file.
